Updated manager timing, and sorting example
Updated manager to handle timing little differently. Now in do-until loop it just checks whether time is met. This is slightly less calculations (though in grand pictures maybe not that important, who knows). More importantly it now tries to dynamically adjust the margin, if frame-budget is exceeded constantly. This is because you really cannot time GPU usage, so although CPU is fine, frame might require to wait for GPU. If you are doing lot of GPU calculations, this might cause constant lag and drop of FPS. This can't be timed nicely within frame, so after the frame margin is adjusted for future frames using delta-time.

Also I updated sorting example.